scraper/requirements.txt
httpx==0.27.0
supabase==2.3.4
python-dotenv==1.0.1
